refactor(toast): file organization
