+++ packages/admin-ui/src/toast/use-toast-queue-state.ts
@@ -1,0 +1,100 @@
+import { useReducer, useCallback } from 'react'
+import type { InternalToast } from './types'
+
+export function useToastQueueState() {
+  const [{ toasts }, dispatch] = useReducer(queueReducer, {
+    toasts: [],
+    queue: {},
+  })
+
+  const add = useCallback(
+    (toast: InternalToast) => dispatch({ type: 'enqueue', toast }),
+    []
+  )
+
+  const remove = useCallback(
+    (key: string) => dispatch({ type: 'dequeue', key }),
+    []
+  )
+
+  return {
+    toasts,
+    add,
+    remove,
+  }
+}
+
+function queueReducer(
+  state: ToastQueueState,
+  action: QueueActions
+): ToastQueueState {
+  switch (action.type) {
+    case 'enqueue': {
+      const { toast } = action
+
+      const hasExistingToast = state.toasts.some(
+        (t) => t.dedupeKey === toast.dedupeKey
+      )
+
+      if (hasExistingToast) {
+        const { dedupeKey } = toast
+
+        return {
+          toasts: state.toasts.map((t) => {
+            if (t.dedupeKey === dedupeKey) {
+              return {
+                ...t,
+                shouldRemove: true,
+              }
+            }
+
+            return t
+          }),
+          queue: {
+            ...state.queue,
+            [dedupeKey]: toast,
+          },
+        }
+      }
+
+      return {
+        ...state,
+        toasts: [...state.toasts, toast],
+      }
+    }
+
+    case 'dequeue': {
+      const toasts = state.toasts.filter(
+        ({ dedupeKey }) => dedupeKey !== action.key
+      )
+
+      const queuedToast = state.queue[action.key]
+
+      if (queuedToast) {
+        return {
+          queue: {
+            ...state.queue,
+            [action.key]: undefined,
+          },
+          toasts: [...toasts, queuedToast],
+        }
+      }
+
+      return {
+        ...state,
+        toasts,
+      }
+    }
+  }
+}
+
+type QueueActions =
+  | { type: 'enqueue'; toast: InternalToast }
+  | { type: 'dequeue'; key: string }
+
+export interface ToastQueueState {
+  toasts: InternalToast[]
+  queue: {
+    [key: string]: InternalToast | undefined
+  }
+}
